+++ lib/location/location_event.dart
@@ -1,0 +1,13 @@
+import 'package:sendrax/models/climb.dart';
+
+abstract class LocationEvent {}
+
+class ClearClimbsEvent extends LocationEvent {}
+
+class ClimbsUpdatedEvent extends LocationEvent {
+  ClimbsUpdatedEvent(this.climbs);
+
+  final List<Climb> climbs;
+}
+
+class LocationErrorEvent extends LocationEvent {}