+++ lib/main/main_state.dart
@@ -9,13 +9,13 @@ class MainState {
   MainState._internal(this.isLoading, this.locations,
       {this.loggedIn = true, this.selected});
 
-  factory MainState.initial() => MainState._internal(false, List<Location>(0));
+  factory MainState.initial() => MainState._internal(false, <Location>[]);
 
   factory MainState.isLoading(bool isLoading, MainState state) =>
       MainState._internal(isLoading, state.locations);
 
-  factory MainState.locations(List<Location> location, MainState state) =>
-      MainState._internal(state.isLoading, location);
+  factory MainState.locations(List<Location> locations, MainState state) =>
+      MainState._internal(state.isLoading, locations);
 
   factory MainState.openLocation(SelectedLocation location, MainState state) =>
       MainState._internal(false, state.locations, selected: location);