+++ lib/models/climb.dart
@@ -1,9 +1,10 @@
 import 'attempt.dart';
 
 class Climb {
-  Climb(this.grade, this.section, [this.attempts]);
+  Climb(this.grade, this.section, this.categories, [this.attempts]);
 
   final String grade;
   final String section;
+  final List<String> categories;
   final List<Attempt> attempts;
 }