+++ lib/models/location.dart
@@ -8,9 +8,12 @@ class SelectedLocation {
 }
 
 class Location {
-  Location(this.id, this.displayName, [this.climbs]);
+  Location(this.id, this.displayName, this.grades,
+      [this.sections, this.climbs]);
 
   final String id;
   final String displayName;
+  final List<String> grades;
+  final List<String> sections;
   final List<Climb> climbs;
 }