+++ lib/navigation_helper.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import 'location/location_view.dart';
 import 'login/login_view.dart';
 import 'main/main_view.dart';
 
@@ -23,6 +24,25 @@ class NavigationHelper {
     } else {
       Navigator.pushReplacement(
           context, MaterialPageRoute(builder: (context) => MainScreen()));
+    }
+  }
+
+  static void navigateToLocation(BuildContext context,
+      String displayName,
+      String locationId,
+      { bool addToBackStack: false }) {
+    if (addToBackStack) {
+      Navigator.push(
+          context,
+          MaterialPageRoute(builder: (context) =>
+              LocationScreen(displayName: displayName, locationId: locationId))
+      );
+    } else {
+      Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) =>
+              LocationScreen(displayName: displayName, locationId: locationId))
+      );
     }
   }
 }